house keeping, remove unnecessary changes
