+++ crates/ironrdp-client/src/lib.rs
@@ -13,5 +13,5 @@ extern crate tracing;
 pub mod clipboard;
 pub mod config;
 pub mod gui;
-mod network_cllient;
 pub mod rdp;
+pub mod network_client;
