+++ crates/ironrdp-web/Cargo.toml
@@ -31,7 +31,7 @@ ironrdp-rdcleanpath.workspace = true
 # WASM
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
-web-sys = { version = "0.3", features = ["HtmlCanvasElement", "ReadableStreamDefaultReader"] }
+web-sys = { version = "0.3", features = ["HtmlCanvasElement","ReadableStreamDefaultReader"] }
 js-sys = "0.3"
 gloo-net = { version = "0.4", default-features = false, features = ["websocket", "http"] }
 gloo-timers = { version = "0.3", default-features = false, features = ["futures"] }
